src: renamed r* archiver modules to repo_*, fixed imports
